fix: Dexquiz/Configurate build: revert bumping

+++ src/main/kotlin/de/tectoast/emolga/utils/json/Emolga.kt
@@ -18,7 +18,7 @@ class Emolga(
     val movefilter: Map<String, List<String>>,
     val mutedroles: Map<Long, Long>,
     val customcommands: MutableMap<String, CCData>,
-    val configuration: MutableMap<Long, MutableMap<String, MutableMap<String, Long>>>,
+    val configuration: MutableMap<Long, MutableMap<String, MutableMap<String, Int>>>,
     val asls11nametoid: List<Long>
 ) {
     fun league(name: String) = drafts[name]!!
